backend/pyproject.toml
[project]
name = "lab-pilot-backend"
version = "0.1.0"
description = "LAB Pilot Demo — FastAPI backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sila2==0.14.0",
    "websockets>=12.0",
    # 인증
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.2.0,<4.1.0",
    # DB
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # 설정
    "pydantic-settings>=2.0.0",
    "pydantic[email]>=2.0.0",
]

[project.scripts]
lab-pilot-backend = "app.main:run"

[tool.setuptools.packages.find]
include = ["app*"]
